// src/app/models/Animal.ts
export default class Animal {
    id: number;
    numeroAnimal:string;
    datanasc:Date;
    produtorid: number; 
    nomeprodutor: string; 
    fazendaid: number; 
    nomefazenda: string; 
    categoriaid: number; 
    nomecategoria: string; 
    racaid: number; 
    nomeraca: string; 
    ativo: boolean;     
  }