// src/app/models/Manejo.ts
export default class Manejo {
    id: number;
    datamanejo: Date; 
    tipomanejoid: number; 
    produtorid: number; 
    fazendaid: number; 
    animalid: number; 
    numeroanimal: string; 
    categoriaid: number; 
    racaid: number; 
    peso: number; 
    aparteid: number;    
    obs: string;  
  }